+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm
 
+## 1.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.2
+
 ## 1.5.1
 
 ### Patch Changes